adding sync on messageBus , starting to parse json at main

+++ src/main/java/bgu/spl/mics/application/objects/CloudPoint.java
@@ -7,8 +7,8 @@ package bgu.spl.mics.application.objects;
 public class CloudPoint {
 
     // TODO: Define fields and methods.
-    private final int x;
-    private final int y;
+    private final double x;
+    private final double y;
 
     public CloudPoint(int x, int y) {
         this.x = x;
@@ -17,11 +17,11 @@ public class CloudPoint {
 
     // Getters and setters (if needed)
     
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
